Smoke looks a little better again through giving the particles a random orientation and making darker smoke less transparent. Removed multi-threading from smoke renderer.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@754 01821946-851b-0410-b7f4-f1477c70e57a

+++ 2d/font.c
@@ -1118,10 +1118,10 @@ void grs_font_read (grs_font *gf, CFILE *fp)
 	gf->ft_minchar = CFReadByte (fp);
 	gf->ft_maxchar = CFReadByte (fp);
 	gf->ft_bytewidth = CFReadShort (fp);
-	gf->ft_data = (ubyte *) ((long) CFReadInt (fp));
-	gf->ft_chars = (ubyte **) ((long) CFReadInt (fp));
-	gf->ft_widths = (short *) ((long) CFReadInt (fp));
-	gf->ft_kerndata = (ubyte *) ((long) CFReadInt (fp));
+	gf->ft_data = (ubyte *) (size_t) CFReadInt (fp);
+	gf->ft_chars = (ubyte **) (size_t) CFReadInt (fp);
+	gf->ft_widths = (short *) (size_t) CFReadInt (fp);
+	gf->ft_kerndata = (ubyte *) (size_t) CFReadInt (fp);
 }
 #endif
